git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@5421 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/IEventReceiver.h
@@ -536,10 +536,10 @@ struct SEvent
 	struct SUserEvent
 	{
 		//! Some user specified data as int
-		s32 UserData1;
+		size_t UserData1;
 
 		//! Another user specified data as int
-		s32 UserData2;
+		size_t UserData2;
 	};
 
 	// Raw events from the OS
